split scarb-execute into subfiles
